backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
requests
python-dotenv
alembic
openai
numpy
pandas
Pillow

